% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_deliver.R
\name{pw_deliver}
\alias{pw_deliver}
\title{Scrape using Readability.js}
\usage{
pw_deliver(url, type = c("static", "dynamic"))
}
\arguments{
\item{url}{The URL of the webpage to extract the article from.}

\item{type}{either "static" or "dynamic". Dynamic uses puppeteer}
}
\value{
The article content as a list (title and content).
}
\description{
Scrape using Readability.js
}
