% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_deliver.R
\name{pw_deliver}
\alias{pw_deliver}
\title{Scrape using Readability.js}
\usage{
pw_deliver(x, type = c("static", "dynamic"))
}
\arguments{
\item{x}{Either a vector of urls or a data.frame returned by \code{\link[paperboy:pb_collect]{paperboy::pb_collect()}}.}

\item{type}{either "static" or "dynamic" if articles are scraped}
}
\value{
A tibble similar to the output of \code{\link[paperboy:pb_deliver]{paperboy::pb_deliver()}}.
}
\description{
Scrape using Readability.js
}
