% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_deliver.R
\name{pw_deliver}
\alias{pw_deliver}
\title{Scrape using Readability.js}
\usage{
pw_deliver(x, type = c("static", "dynamic"))
}
\arguments{
\item{x}{Either a url or a data.frame returned by \link{paperboy::pb_collect}.}

\item{type}{either "static" or "dynamic" if articles are scraped}
}
\value{
The article content as a list (title and content).
}
\description{
Scrape using Readability.js
}
