% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pw_report}
\alias{pw_report}
\title{Summary of delivered articles}
\usage{
pw_report(x, n = 100)
}
\arguments{
\item{x}{result from \code{\link[=pw_deliver]{pw_deliver()}}}

\item{n}{integer cutoff when articles are considered too short (default 100)}
}
\value{
nothing. called for side effects
}
\description{
Summary of delivered articles
}
